Split graph tests in files and subtests
