+++ nucliadb/tests/nucliadb/integration/test_relations.py
@@ -20,9 +20,6 @@
 import pytest
 from httpx import AsyncClient
 
-from nucliadb_models.graph import responses as graph_responses
-from nucliadb_models.graph.responses import GraphNodesSearchResponse, GraphSearchResponse
-from nucliadb_models.metadata import RelationEntity, RelationNodeType
 from nucliadb_protos.resources_pb2 import (
     FieldComputedMetadataWrapper,
     FieldType,
@@ -248,467 +245,3 @@ async def create_broker_message_with_relations():
     bm.field_metadata.append(fcmw)
 
     return bm
-
-
-@pytest.fixture(scope="function")
-def knowledge_graph() -> tuple[dict[str, RelationEntity], list[tuple[str, str, str]]]:
-    entities = {
-        "Anastasia": RelationEntity(value="Anastasia", type=RelationNodeType.ENTITY, group="PERSON"),
-        "Anna": RelationEntity(value="Anna", type=RelationNodeType.ENTITY, group="PERSON"),
-        "Apollo": RelationEntity(value="Apollo", type=RelationNodeType.ENTITY, group="PROJECT"),
-        "Cat": RelationEntity(value="Cat", type=RelationNodeType.ENTITY, group="ANIMAL"),
-        "Climbing": RelationEntity(value="Climbing", type=RelationNodeType.ENTITY, group="ACTIVITY"),
-        "Computer science": RelationEntity(
-            value="Computer science", type=RelationNodeType.ENTITY, group="STUDY_FIELD"
-        ),
-        "Dimitri": RelationEntity(value="Dimitri", type=RelationNodeType.ENTITY, group="PERSON"),
-        "Erin": RelationEntity(value="Erin", type=RelationNodeType.ENTITY, group="PERSON"),
-        "Jerry": RelationEntity(value="Jerry", type=RelationNodeType.ENTITY, group="ANIMAL"),
-        "Margaret": RelationEntity(value="Margaret", type=RelationNodeType.ENTITY, group="PERSON"),
-        "Mouse": RelationEntity(value="Mouse", type=RelationNodeType.ENTITY, group="ANIMAL"),
-        "New York": RelationEntity(value="New York", type=RelationNodeType.ENTITY, group="PLACE"),
-        "Olympic athlete": RelationEntity(
-            value="Olympic athlete", type=RelationNodeType.ENTITY, group="SPORT"
-        ),
-        "Peter": RelationEntity(value="Peter", type=RelationNodeType.ENTITY, group="PERSON"),
-        "Rocket": RelationEntity(value="Rocket", type=RelationNodeType.ENTITY, group="VEHICLE"),
-        "Tom": RelationEntity(value="Tom", type=RelationNodeType.ENTITY, group="ANIMAL"),
-        "UK": RelationEntity(value="UK", type=RelationNodeType.ENTITY, group="PLACE"),
-    }
-    graph = [
-        ("Anastasia", "IS_FRIEND", "Anna"),
-        ("Anna", "FOLLOW", "Erin"),
-        ("Anna", "LIVE_IN", "New York"),
-        ("Anna", "LOVE", "Cat"),
-        ("Anna", "WORK_IN", "New York"),
-        ("Apollo", "IS", "Rocket"),
-        ("Dimitri", "LOVE", "Anastasia"),
-        ("Erin", "BORN_IN", "UK"),
-        ("Erin", "IS", "Olympic athlete"),
-        ("Erin", "LOVE", "Climbing"),
-        ("Jerry", "IS", "Mouse"),
-        ("Margaret", "DEVELOPED", "Apollo"),
-        ("Margaret", "WORK_IN", "Computer science"),
-        ("Peter", "LIVE_IN", "New York"),
-        ("Tom", "CHASE", "Jerry"),
-        ("Tom", "IS", "Cat"),
-    ]
-
-    return (entities, graph)
-
-
-@pytest.mark.deploy_modes("standalone")
-async def test_user_defined_knowledge_graph(
-    nucliadb_reader: AsyncClient,
-    nucliadb_writer: AsyncClient,
-    standalone_knowledgebox: str,
-    knowledge_graph: tuple[dict[str, RelationEntity], list[tuple[str, str, str]]],
-):
-    kbid = standalone_knowledgebox
-    entities, paths = knowledge_graph
-
-    graph = paths[:3]
-    resp = await nucliadb_writer.post(
-        f"/kb/{kbid}/resources",
-        json={
-            "title": "Knowledge graph",
-            "slug": "knowledge-graph",
-            "summary": "User defined knowledge graph",
-            "usermetadata": {
-                "relations": [
-                    {
-                        "relation": "ENTITY",
-                        "label": relation,
-                        "from": entities[source].model_dump(),
-                        "to": entities[target].model_dump(),
-                    }
-                    for source, relation, target in graph
-                ],
-            },
-        },
-    )
-    assert resp.status_code == 201
-    rid = resp.json()["uuid"]
-
-    resp = await nucliadb_reader.get(f"/kb/{kbid}/resource/{rid}?show=basic")
-    assert resp.status_code == 200
-    user_graph = resp.json()["usermetadata"]["relations"]
-    assert len(user_graph) == len(graph)
-
-    # Update graph
-
-    graph = paths
-    resp = await nucliadb_writer.patch(
-        f"/kb/{kbid}/resource/{rid}",
-        json={
-            "usermetadata": {
-                "relations": [
-                    {
-                        "relation": "ENTITY",
-                        "label": relation,
-                        "from": entities[source].model_dump(),
-                        "to": entities[target].model_dump(),
-                    }
-                    for source, relation, target in graph
-                ],
-            },
-        },
-    )
-    assert resp.status_code == 200
-
-    resp = await nucliadb_reader.get(f"/kb/{kbid}/resource/{rid}?show=basic")
-    assert resp.status_code == 200
-    user_graph = resp.json()["usermetadata"]["relations"]
-    assert len(user_graph) == len(graph)
-
-    # Search graph
-
-    resp = await nucliadb_reader.post(
-        f"/kb/{kbid}/find",
-        json={
-            "query_entities": [
-                {
-                    "name": entities["Anna"].value,
-                    "type": entities["Anna"].type.value,
-                    "subtype": entities["Anna"].group,
-                }
-            ],
-            "features": ["relations"],
-        },
-    )
-    assert resp.status_code == 200
-    body = resp.json()
-    retrieved_graph = set()
-    for source, subgraph in body["relations"]["entities"].items():
-        for target in subgraph["related_to"]:
-            if target["direction"] == "in":
-                path = (source, "-", target["relation_label"], "->", target["entity"])
-            else:
-                path = (source, "<-", target["relation_label"], "-", target["entity"])
-
-            assert path not in retrieved_graph, "We don't expect duplicated paths"
-            retrieved_graph.add(path)
-
-    for relation in retrieved_graph:
-        print(relation)
-
-    assert len(retrieved_graph) == 5
-
-
-@pytest.mark.deploy_modes("standalone")
-async def test_graph_search(
-    nucliadb_reader: AsyncClient,
-    nucliadb_writer: AsyncClient,
-    standalone_knowledgebox: str,
-    knowledge_graph: tuple[dict[str, RelationEntity], list[tuple[str, str, str]]],
-):
-    kbid = standalone_knowledgebox
-    entities, paths = knowledge_graph
-
-    resp = await nucliadb_writer.post(
-        f"/kb/{kbid}/resources",
-        json={
-            "usermetadata": {
-                "relations": [
-                    {
-                        "relation": "ENTITY",
-                        "label": relation,
-                        "from": entities[source].model_dump(),
-                        "to": entities[target].model_dump(),
-                    }
-                    for source, relation, target in paths
-                ],
-            },
-        },
-    )
-    assert resp.status_code == 201
-
-    def simple_paths(paths: list[graph_responses.GraphPath]) -> list[tuple[str, str, str]]:
-        simple_paths = []
-        for path in paths:
-            # response should never return empty nodes/relations
-            assert path.source is not None
-            assert path.source.value is not None
-            assert path.relation is not None
-            assert path.relation.label is not None
-            assert path.destination is not None
-            assert path.destination.value is not None
-            simple_paths.append((path.source.value, path.relation.label, path.destination.value))
-        return simple_paths
-
-    ## Node queries
-
-    # ()-[]->()
-    resp = await nucliadb_reader.post(
-        f"/kb/{kbid}/graph",
-        json={
-            "query": {
-                "source": {},
-            },
-            "top_k": 100,
-        },
-    )
-    assert resp.status_code == 200
-    paths = simple_paths(GraphSearchResponse.model_validate(resp.json()).paths)
-    assert len(paths) == 16
-
-    # (:PERSON)-[]->()
-    resp = await nucliadb_reader.post(
-        f"/kb/{kbid}/graph",
-        json={
-            "query": {
-                "source": {
-                    "group": "PERSON",
-                },
-            },
-            "top_k": 100,
-        },
-    )
-    assert resp.status_code == 200
-    paths = simple_paths(GraphSearchResponse.model_validate(resp.json()).paths)
-    assert len(paths) == 12
-
-    # (:Anna)-[]->()
-    resp = await nucliadb_reader.post(
-        f"/kb/{kbid}/graph",
-        json={
-            "query": {
-                "source": {
-                    "value": "Anna",
-                    "group": "PERSON",
-                },
-            },
-            "top_k": 100,
-        },
-    )
-    assert resp.status_code == 200
-    paths = simple_paths(GraphSearchResponse.model_validate(resp.json()).paths)
-    assert len(paths) == 4
-    assert ("Anna", "FOLLOW", "Erin") in paths
-    assert ("Anna", "LIVE_IN", "New York") in paths
-    assert ("Anna", "LOVE", "Cat") in paths
-    assert ("Anna", "WORK_IN", "New York") in paths
-
-    # ()-[]->(:Anna)
-    resp = await nucliadb_reader.post(
-        f"/kb/{kbid}/graph",
-        json={
-            "query": {
-                "destination": {
-                    "value": "Anna",
-                    "group": "PERSON",
-                },
-            },
-            "top_k": 100,
-        },
-    )
-    assert resp.status_code == 200
-    paths = simple_paths(GraphSearchResponse.model_validate(resp.json()).paths)
-    assert len(paths) == 1
-    assert ("Anastasia", "IS_FRIEND", "Anna") in paths
-
-    # (:Anna)
-    resp = await nucliadb_reader.post(
-        f"/kb/{kbid}/graph",
-        json={
-            "query": {
-                "source": {
-                    "value": "Anna",
-                    "group": "PERSON",
-                },
-                "undirected": True,
-            },
-            "top_k": 100,
-        },
-    )
-    assert resp.status_code == 200
-    paths = simple_paths(GraphSearchResponse.model_validate(resp.json()).paths)
-    assert len(paths) == 5
-    assert ("Anna", "FOLLOW", "Erin") in paths
-    assert ("Anna", "LIVE_IN", "New York") in paths
-    assert ("Anna", "LOVE", "Cat") in paths
-    assert ("Anna", "WORK_IN", "New York") in paths
-    assert ("Anastasia", "IS_FRIEND", "Anna") in paths
-
-    ## Fuzzy node queries
-
-    # (:Anna)
-    resp = await nucliadb_reader.post(
-        f"/kb/{kbid}/graph",
-        json={
-            "query": {
-                "source": {
-                    "value": "Anna",
-                    "group": "PERSON",
-                },
-                "undirected": True,
-            },
-            "top_k": 100,
-        },
-    )
-    assert resp.status_code == 200
-    paths = simple_paths(GraphSearchResponse.model_validate(resp.json()).paths)
-    assert len(paths) == 5
-    assert ("Anna", "FOLLOW", "Erin") in paths
-    assert ("Anna", "LIVE_IN", "New York") in paths
-    assert ("Anna", "LOVE", "Cat") in paths
-    assert ("Anna", "WORK_IN", "New York") in paths
-    assert ("Anastasia", "IS_FRIEND", "Anna") in paths
-
-    ## Relations
-
-    # ()-[LIVE_IN]->()
-    resp = await nucliadb_reader.post(
-        f"/kb/{kbid}/graph",
-        json={
-            "query": {
-                "relation": {
-                    "label": "LIVE_IN",
-                },
-            },
-            "top_k": 100,
-        },
-    )
-    assert resp.status_code == 200
-    paths = simple_paths(GraphSearchResponse.model_validate(resp.json()).paths)
-    assert len(paths) == 2
-    assert ("Anna", "LIVE_IN", "New York") in paths
-    assert ("Peter", "LIVE_IN", "New York") in paths
-
-    ## Directed paths
-
-    # (:Erin)-[]->(:UK)
-    resp = await nucliadb_reader.post(
-        f"/kb/{kbid}/graph",
-        json={
-            "query": {
-                "source": {
-                    "value": "Erin",
-                    "group": "PERSON",
-                },
-                "destination": {
-                    "value": "UK",
-                    "group": "PLACE",
-                },
-            },
-            "top_k": 100,
-        },
-    )
-    assert resp.status_code == 200
-    paths = simple_paths(GraphSearchResponse.model_validate(resp.json()).paths)
-    assert len(paths) == 1
-    assert ("Erin", "BORN_IN", "UK") in paths
-
-    # (:PERSON)-[]->(:PLACE)
-    resp = await nucliadb_reader.post(
-        f"/kb/{kbid}/graph",
-        json={
-            "query": {
-                "source": {
-                    "group": "PERSON",
-                },
-                "destination": {
-                    "group": "PLACE",
-                },
-            },
-            "top_k": 100,
-        },
-    )
-    assert resp.status_code == 200
-    paths = simple_paths(GraphSearchResponse.model_validate(resp.json()).paths)
-    assert len(paths) == 4
-    assert ("Anna", "LIVE_IN", "New York") in paths
-    assert ("Anna", "WORK_IN", "New York") in paths
-    assert ("Erin", "BORN_IN", "UK") in paths
-    assert ("Peter", "LIVE_IN", "New York") in paths
-
-    # (:PERSON)-[LIVE_IN]->(:PLACE)
-    resp = await nucliadb_reader.post(
-        f"/kb/{kbid}/graph",
-        json={
-            "query": {
-                "source": {
-                    "group": "PERSON",
-                },
-                "relation": {
-                    "label": "LIVE_IN",
-                },
-                "destination": {
-                    "group": "PLACE",
-                },
-            },
-            "top_k": 100,
-        },
-    )
-    assert resp.status_code == 200
-    paths = simple_paths(GraphSearchResponse.model_validate(resp.json()).paths)
-    assert len(paths) == 2
-    assert ("Anna", "LIVE_IN", "New York") in paths
-    assert ("Peter", "LIVE_IN", "New York") in paths
-
-    ## Undirected paths
-    # (:Anna)-[:IS_FRIEND]-()
-    resp = await nucliadb_reader.post(
-        f"/kb/{kbid}/graph",
-        json={
-            "query": {
-                "source": {
-                    "value": "Anna",
-                },
-                "relation": {
-                    "label": "IS_FRIEND",
-                },
-                "undirected": True,
-            },
-            "top_k": 100,
-        },
-    )
-    assert resp.status_code == 200
-    paths = simple_paths(GraphSearchResponse.model_validate(resp.json()).paths)
-    assert len(paths) == 1
-    assert ("Anastasia", "IS_FRIEND", "Anna") in paths
-
-
-@pytest.mark.skip(reason="Early development")
-@pytest.mark.deploy_modes("standalone")
-async def test_graph_nodes_search(
-    nucliadb_reader: AsyncClient,
-    nucliadb_writer: AsyncClient,
-    standalone_knowledgebox: str,
-    knowledge_graph: tuple[dict[str, RelationEntity], list[tuple[str, str, str]]],
-):
-    kbid = standalone_knowledgebox
-    entities, paths = knowledge_graph
-
-    resp = await nucliadb_writer.post(
-        f"/kb/{kbid}/resources",
-        json={
-            "usermetadata": {
-                "relations": [
-                    {
-                        "relation": "ENTITY",
-                        "label": relation,
-                        "from": entities[source].model_dump(),
-                        "to": entities[target].model_dump(),
-                    }
-                    for source, relation, target in paths
-                ],
-            },
-        },
-    )
-    assert resp.status_code == 201
-
-    # (:PERSON)-[]->()
-    resp = await nucliadb_reader.post(
-        f"/kb/{kbid}/graph/nodes",
-        json={
-            "query": {
-                "position": "source",
-                "group": "PERSON",
-            },
-            "top_k": 100,
-        },
-    )
-    assert resp.status_code == 200
-    nodes = GraphNodesSearchResponse.model_validate(resp.json()).nodes
-    assert len(nodes) == 12
